backend/petpal/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
Pillow
drf-yasg
django-extensions
django
django-cors-headers
django-environ
psycopg2-binary
